make validando pacotes automaticamente
