+++ {{ cookiecutter.project_slug }}/Makefile
@@ -1,16 +1,23 @@
-.PHONY: all extract validate transform build check publish clean
+.PHONY: requirements all extract validate transform build check publish clean
+
 
 EXT = txt
-CHECK_PKG := $(shell python scripts/check.py)
-RESOURCE_NAMES := $(shell python main.py resources)
+
+RESOURCE_NAMES := $(shell python main.py resources) && exit 1
 OUTPUT_FILES := $(addsuffix .csv,$(addprefix data/,$(RESOURCE_NAMES)))
 
-all: extract validate transform build check
+all: requirements extract validate transform build check
 
-extract: 
+%: requirements
+	:
+
+requirements:
+	python main.py requirements
+
+extract:
 	$(foreach resource_name, $(RESOURCE_NAMES),python main.py extract $(resource_name) &&) true
 
-validate: 
+validate:
 	frictionless validate datapackage.yaml
 
 transform: $(OUTPUT_FILES)
@@ -24,9 +31,9 @@ datapackage.json: $(OUTPUT_FILES) scripts/build.py datapackage.yaml
 	python main.py build
 
 check:
-	frictionless validate datapackage.json
+	frictionless validate datapackage.yaml
 
-publish: 
+publish:
 	git add -Af datapackage.json data/*.csv data-raw/*.$(EXT)
 	git commit --author="Automated <actions@users.noreply.github.com>" -m "Update data package at: $$(date +%Y-%m-%dT%H:%M:%SZ)" || exit 0
 	git push
